talipot-core: Remove some dreadful STL hacks
Hacking on the STL internals is not a great idea for software maintenance.

So remove those std::vector hacks to avoid calling constructors when resizing,
stick to the STL API and let the C++ compiler do its job to optimize the code.

Perormance gains are once again negligible in Release mode with those type
of hacks and it makes the code awful to read and understand.