talipot-python/FindReplaceDialog: Add parent widget for better GUI behavior
Ensure to set a parent widget to the "Find / Replace" dialog to synchronize visibility
state with its top level window (main window in anchored mode, IDE window in non
anchored mode) and avoid having the dialog referenced in the OS task bar.

After that change, fix a display bug due to the Qt stylesheets set on the dialog:
Python IDE in non anchored mode was not repainted when trying to display it again
after closing it.

Reinitialize the "Find / Replace" dialogs when anchoring / unanchoring the Python
IDE in order to set proper parent widget (main window or IDE dialog).

Hide the "Find / Replace" dialog when the corresponding code editor gets hidden.