gui: Replace outdated GUI bitmap icons by Material Design ones
The GUI bitmap icons set inherited from the Tulip era was looking so outdated
and not really consistent in terms of design.

So replace all those bitmap icons by Material Design ones as it enables to
get high resolution icons in a consistent and nice style.

The icons are generated from the Material Design Icons webfont which allows
to get icons with every possible colors, scales or geometric transformations.

To respect the flat look GUI conventions, a few colors are used to generate
the icons to get a nice and pleasant display.