Remove the use of doxygen macros in source code
Future documentation for Talipot C++ API will be generated by Sphinx and the Breathe
extension so remove doxygen output filtering through the use of macros in the source
code as we do not care about it anymore (and it adds a lot of noise in sources).

Also remove no more used doxygen files from the Tulip era.