plugins/layout/OGDF: Workaround possible segfaults when loading plugins
A weird segfault in OGDF code was observed when running the Talipot AppImage
on Ubuntu 20.04 and plugins are loaded (but not on Debian stable or ArchLinux).

Workaround the issue by instantiating OGDF algorithms only when a non null
plugin context is provided (a null context is passed when plugins are loaded).