talipot-gui/TlpQtTools: Get GUI base colors from QApplication palette
By proceeding like this, Talipot look and feel can be adapted to every desktop
appearances (KDE or MacOS dark theme for instance).

For GTK based desktops, qt5ct should be used to configure Talipot appearance
in the same way as the desktop theme.

Also detect if application is in dark mode by checking the lightness value
of the background color.