plugins/interactor/Fisheye: Move towards an image-based implementation
Instead of implementing the fisheye distortion using an OpenGL vertex
shader, achieve the same effect with rendering to texture and an OpenGL
fragment shader.

The area of the OpenGL scene to magnify will be rendered with higher
details in a larger framebuffer object and the fisheye distortion will
then be applied by a fragment shader using the framebuffer texture as
input.

The advantage of proceeding like this is that the fisheye distortion
can be applied to any OpenGL scene without the need to modify the
input vertices geometry.