plugins/OGDFFm3: Remove some automatically set parameters
As high level options for the FMMMLayout are used, the number of fixed
iterations and the initial placement forces will be automatically set
based on the "Quality vs Speed" parameter value or the
"New initial placement" value.

So remove the "Fixed iterations" and "Initial Placement Forces"
parameters as their value will not be taken into account.