talipot-ogl/GlLabel: Add preliminary support for text shaping
Use the fribidi library to execute Unicode Bidirectional algorithm
on the text to render.

It enables to improve the rendering of complex text layout like
arabic scripts.

More complex scripts will require to use the harfbuzz library for
better text shaping but this will be done when refactoring the
OpenGL rendering engine.