talipot-core/GraphView: Simplify and improve implementation
Store the edges incidence list for each node in the GraphView class
used to implement subgraphs. Nevertheless the edges extremities remain
stored in the root graph of a hierarchy implemented in GraphImpl class.

It enables to significantly speed up iterations on node adjancencies
and incidence for subgraphs as those are no more filtered from the
root graph ones.

It also means that edge ordering is now local to each subgraph.

Rename allEdges method from the Graph interface to incidence.

Simplify and cleanup a lot of code using iterator helpers in order to
improve readability.